+++ MHLibDemos/src/MKP.jl
@@ -8,15 +8,11 @@ Each item has a price and requires from each resource a certain amount.
 Find a subset of the items with maximum total price that does not exceed the resources'
 capacities.
 """
-module MKP
 
 using ArgParse
-
 using MHLib
-using MHLib.Schedulers
-using MHLib.SubsetVectorSolutions
 
-export MKPInstance, MKPSolution
+export MKPInstance, MKPSolution, solve_mkp
 
 """
     MKPInstance
@@ -135,7 +131,7 @@ function MHLib.check(s::MKPSolution; kwargs...)
     y_old = s.y
     calc_y!(s)
     if any(y_old .!= s.y)
-        error("Solution had invalid y values: $(s.y) $(s.y_old)")
+        error("Solution had invalid y values: $(s.y) $(y_old)")
     end
     if any(s.y .> s.inst.b)
         error("Solution exceeds capacity limits: $(self.y) $(s.inst.b)")
@@ -213,4 +209,39 @@ function MHLib.SubsetVectorSolutions.element_added_delta_eval!(s::MKPSolution;
     return false
 end
 
-end # module
+# -------------------------------------------------------------------------------
+
+function solve_mkp(args=ARGS)
+    println("MKP Demo version $(git_version())\nARGS: ", args)
+
+    # set some new default values for parameters and parse all relevant arguments
+    settings_new_default_value!(MHLib.settings_cfg, "ifile", "data/mknapcb5-01.txt")
+    settings_new_default_value!(MHLib.Schedulers.settings_cfg, "mh_titer", 5000)
+    parse_settings!([MHLib.Schedulers.settings_cfg], args)
+    println(get_settings_as_string())
+
+    inst = MKPInstance(settings[:ifile])
+    sol = MKPSolution(inst)
+    # initialize!(sol)
+    # check(sol)
+    # println(sol)
+
+    # we apply a variable neighborhood search:
+    alg = GVNS(sol, [MHMethod("con", construct!, 0)],
+        [MHMethod("li1", local_improve!, 1)],
+        [MHMethod("sh1", shaking!, 1), MHMethod("sh2", shaking!, 2),
+            MHMethod("sh3", shaking!, 3)], 
+        consider_initial_sol = true)
+    run!(alg)
+    method_statistics(alg.scheduler)
+    main_results(alg.scheduler)
+    check(sol)
+    return sol
+end
+
+# To run from REPL, use MHLibDemos and call `solve_mkp(<args>)` where `<args>` is 
+# a list of strings being passed as arguments for setting global parameters.
+# `@<filename>` may be used to read arguments from a configuration file <filename>
+
+# Run with profiler:
+# @profview solve_mkp(args)