+++ MHLibDemos/src/OneMax.jl
@@ -1,0 +1,86 @@
+"""
+    OneMax
+
+OneMax demo problem: Maximize the number of set bits in a binary string.
+
+This problem is just for simple demonstration/debugging purposes.
+"""
+
+using ArgParse
+using MHLib
+
+export OneMaxSolution, onemax_settings_cfg, solve_onemax
+
+# We define an additional problem-specific parameter:
+const onemax_settings_cfg = ArgParseSettings()
+@add_arg_table! onemax_settings_cfg begin
+    "--onemax_n"
+        help = "length of solution string in the problem"
+        arg_type = Int
+        default = 100
+end
+
+
+"""
+    OneMaxSolution
+
+A concrete solution type to solve the MAXSAT problem.
+
+As the problem is so simply defined, we do not need a separate instance structure
+but store problem size directly within the solutions.
+"""
+mutable struct OneMaxSolution <: BoolVectorSolution
+    obj_val::Int
+    obj_val_valid::Bool
+    x::Vector{Bool}
+end
+
+OneMaxSolution(n) = OneMaxSolution(-1, false, Vector{Bool}(undef, n))
+
+OneMaxSolution(s::OneMaxSolution) = OneMaxSolution(s.obj_val, s.obj_val_valid, copy(s.x))
+
+MHLib.calc_objective(s::OneMaxSolution) = sum(s.x)
+
+Base.show(io::IO, s::OneMaxSolution) = println(io, s.x)
+
+function Base.copy!(s1::OneMaxSolution, s2::OneMaxSolution)
+    s1.obj_val = s2.obj_val
+    s1.obj_val_valid = s2.obj_val_valid
+    copy!(s1.x, s2.x)
+end
+
+Base.copy(s::OneMaxSolution) = OneMaxSolution(s.obj_val, s.obj_val_valid, copy(s.x))
+
+
+# -------------------------------------------------------------------------------
+
+function solve_onemax(args=ARGS)
+    println("OneMax Demo version $(git_version())\nARGS: ", args)
+
+    # We set some new default values for parameters and parse all relevant arguments
+    settings_new_default_value!(MHLib.Schedulers.settings_cfg, "mh_titer", 100)
+    parse_settings!([MHLib.Schedulers.settings_cfg, onemax_settings_cfg], args)
+    println(get_settings_as_string())
+        
+    sol = OneMaxSolution(settings[:onemax_n])
+    initialize!(sol)
+    println(sol)
+
+    # We apply here a variable neighborhood search, making use of a simple construction
+    # heuristic, a local improvement method, and a shaking method.
+    alg = GVNS(sol, [MHMethod("con", construct!, 0)],
+        [MHMethod("li1", local_improve!, 1)],[MHMethod("sh1", shaking!, 1)], 
+        consider_initial_sol = true)
+    run!(alg)
+    method_statistics(alg.scheduler)
+    main_results(alg.scheduler)
+    check(sol)
+    return sol
+end
+
+# To run from REPL, use `MHLibDemos` and call `solve_onemax(<args>)` where 
+# `<args>` is a list of strings being passed as arguments for setting global parameters.
+# `@<filename>` may be used to read arguments from a configuration file <filename>
+
+# Run with profiler:
+# @profview solve_onemax(args)