+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Major changes in releases:
 
 ## Upcoming
+
+## Version 0.1.8
+- all demo applications refined and moved into an own subpackage `MHLibDemos`
+- all symbols of submodules are now also re-exported by the main `MHLib` module
 - LNS/ALNS refactored, `MethodSelector` introduced to generalize LNS
 - all structures adapted to use type parameters instead of abstract types for elements